Custom CSS for Travian to improve the interface
Add this link to <a href="https://ublockorigin.com">uBlock</a> as a blocklist and the interface will become more compact.

Big accounts have to go up and down the page all the time because the list with villages is very long. That's why I reduced the height of the interface in every why I could think of.

Before you can build troops, you will have to add units as favourite. This way you can choose which troops will be displayed so you don't build the wrong units by accident.